requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
prophet>=1.1.0
gitpython>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0